+++ configs/templates/_vcd.txt
@@ -16,8 +16,8 @@ VCD_SNMP_VARIABLE = need_to_be_configured_by_user
 # PEBCAK documentation for the Wizard and CLI
 VCD_ACCESS_DOC = Host name or IP of the vCloud Director REST interface
 VCD_ACCESS_DEFAULT = 10.0.0.1
-VCD_CREDENTIALS_DOCC = Login name and password for the vCloud Director REST iinterface
-VCD_CREDENTIALS_DEFAULT = happyguy@orgname;password
+VCD_CREDENTIALS_DOC = Login name and password for the vCloud Director REST iinterface
+VCD_CREDENTIALS_DEFAULT = happyguy@orgname:password # need colon instead of semi-colon
 VCD_SECURITY_GROUPS_DOC = Not used with vCloud Director
 VCD_SECURITY_GROUPS_DEFAULT = none
 VCD_KEY_NAME_DOC = Password used by the vCloud Director REST interface. (Again)
@@ -74,7 +74,8 @@ SNMP_COMMUNITY = $VCD_SNMP_COMMUNITY
 SNMP_VARIABLE = $VCD_SNMP_VARIABLE
 # CLONE_TIMEOUT is number of seconds libcloud should poll VCD after issuing clone command before timing out
 CLONE_TIMEOUT = 600
-NETNAME = private
+PROV_NETNAME = private
+RUN_NETNAME = private
 TENANT = default
 SIZE = from_vm_template
 LEAVE_INSTANCE_ON_FAILURE = $False
